Run once per sim instead of report
